fix(editor): Resolve TSR warning for CatchAllComponent
