+++ apps/editor.planx.uk/src/routes/__root.tsx
@@ -6,12 +6,11 @@ import {
 import { TanStackRouterDevtools } from "@tanstack/react-router-devtools";
 import { zodValidator } from "@tanstack/zod-adapter";
 import DelayedLoadingIndicator from "components/DelayedLoadingIndicator/DelayedLoadingIndicator";
+import { CatchAllComponent } from "pages/ErrorPage/CatchAllComponent";
 import ErrorPage from "pages/ErrorPage/ErrorPage";
 import React from "react";
 import { createPortal } from "react-dom";
 import { z } from "zod";
-
-import { CatchAllComponent } from "./$";
 
 // Search params schema for error handling and redirects
 const rootSearchSchema = z.object({
