+++ apps/editor.planx.uk/src/test/utils.tsx
@@ -22,8 +22,8 @@ import userEvent, {
   PointerEventsCheckLevel,
 } from "@testing-library/user-event";
 import { ToastContextProvider } from "contexts/ToastContext";
+import { CatchAllComponent } from "pages/ErrorPage/CatchAllComponent";
 import React from "react";
-import { CatchAllComponent } from "routes/$";
 
 import { defaultTheme } from "../theme";
